Symlink MacVim.app to /Applications.
Homebrew's cask install of MacVim fails because vim is
already installed. The workaround is to install the non-cask
version of MacVim, and then manually symlink the MacVim.app
to /Applications.

For file associations - such as always opening .txt files
with MacVim - the actual application rather than the symlink
must be used, e.g.:

  /usr/local/Cellar/macvim/8.2-166_1/MacVim.app.